third_party/blink/web_tests/external/wpt/css/css-color/parsing/relative-color-out-of-gamut-expected.txt

This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught ReferenceError: test_computed_value is not defined
FAIL e.style['color'] = "rgb(from color(display-p3 0 1 0) r g b / alpha)" should set the property value assert_equals: serialization should be canonical expected "rgb(0, 249, 66)" but got "rgb(0, 255, 0)"
FAIL e.style['color'] = "rgb(from lab(100 104.3 -50.9) r g b)" should set the property value assert_equals: serialization should be canonical expected "rgb(255, 255, 255)" but got "rgb(255, 150, 255)"
FAIL e.style['color'] = "rgb(from lab(0 104.3 -50.9) r g b)" should set the property value assert_equals: serialization should be canonical expected "rgb(42, 0, 34)" but got "rgb(90, 0, 76)"
FAIL e.style['color'] = "rgb(from lch(100 116 334) r g b)" should set the property value assert_equals: serialization should be canonical expected "rgb(255, 255, 255)" but got "rgb(255, 150, 255)"
FAIL e.style['color'] = "rgb(from lch(0 116 334) r g b)" should set the property value assert_equals: serialization should be canonical expected "rgb(42, 0, 34)" but got "rgb(90, 0, 76)"
FAIL e.style['color'] = "rgb(from oklab(1 0.365 -0.16) r g b)" should set the property value assert_equals: serialization should be canonical expected "rgb(255, 255, 255)" but got "rgb(255, 92, 255)"
FAIL e.style['color'] = "rgb(from oklab(0 0.365 -0.16) r g b)" should set the property value assert_equals: serialization should be canonical expected "rgb(0, 0, 0)" but got "rgb(19, 0, 24)"
FAIL e.style['color'] = "rgb(from oklch(1 0.399 336.3) r g b)" should set the property value assert_equals: serialization should be canonical expected "rgb(255, 255, 255)" but got "rgb(255, 91, 255)"
FAIL e.style['color'] = "rgb(from oklch(0 0.399 336.3) r g b)" should set the property value assert_equals: serialization should be canonical expected "rgb(0, 0, 0)" but got "rgb(20, 0, 24)"
Harness: the test ran to completion.
